int HR fix

+++ wear/src/main/java/com/gm/android/DriverStatus/WearActivity.java
@@ -84,7 +84,7 @@ public class WearActivity extends Activity implements HeartbeatService.OnChangeL
 
         switch (Str){
             case "HR":
-                mTextView.setText(Double.toString((int)newValue));
+                mTextView.setText(Integer.toString((int) newValue));
                 break;
             case "AX":
                 mTextView4.setText(Double.toString(newValue));
